作业

+++ src/main/java/cs/sbs/web/model/MenuItem.java
@@ -4,12 +4,16 @@ public class MenuItem {
     private String name;
     private double price;
 
-    public MenuItem(String name, int price) {
+    public MenuItem(String name, double price) {
         this.name = name;
         this.price = price;
     }
 
-    // Getter & Setter
-    public String getName() { return name; }
-    public double getPrice() { return price; }
+    public String getName() {
+        return name;
+    }
+
+    public double getPrice() {
+        return price;
+    }
 }
